Migrate `apps/meteor/client/apps/RealAppsEngineUIHost.js`
